+++ src/ServiceProvider.php
@@ -12,60 +12,7 @@ class ServiceProvider extends LaravelServiceProvider
      *
      * @var bool
      */
-    protected $defer = false;
-
-    /**
-     * Indicates if the application is Laravel 4.
-     *
-     * @var bool
-     */
-    protected $isLaravel4 = false;
-
-    /**
-     * Indicates if the application is Laravel 5.
-     *
-     * @var bool
-     */
-    protected $isLaravel5 = false;
-
-    /**
-     * Indicates if the application is Lumen.
-     *
-     * @var bool
-     */
-    protected $isLumen = false;
-
-    /**
-     * Create a new service provider instance.
-     *
-     * @param  \Illuminate\Contracts\Foundation\Application  $app
-     */
-    public function __construct($app)
-    {
-        parent::__construct($app);
-
-        $appVersion = method_exists($app, 'version') ? $app->version() : $app::VERSION;
-
-        $this->isLaravel4 = (int) $appVersion == 4;
-        $this->isLaravel5 = (int) $appVersion == 5;
-        $this->isLumen = Str::contains($appVersion, 'Lumen');
-    }
-
-    /**
-     * Bootstrap any application services.
-     *
-     * @return void
-     */
-    public function boot()
-    {
-        if ($this->isLaravel4) {
-            $this->package('elfsundae/laravel-bearychat', 'bearychat', __DIR__);
-        } else {
-            $this->publishes([
-                $this->getConfigFromPath() => $this->getConfigToPath(),
-            ], 'laravel-bearychat');
-        }
-    }
+    protected $defer = true;
 
     /**
      * Register the service provider.
@@ -74,68 +21,31 @@ class ServiceProvider extends LaravelServiceProvider
      */
     public function register()
     {
-        if (! $this->isLaravel4) {
-            $this->mergeConfigFrom($this->getConfigFromPath(), 'bearychat');
+        $this->mergeConfigFrom($from = __DIR__.'/../config/bearychat.php', 'bearychat');
+
+        if ($this->app->runningInConsole()) {
+            $to = (function_exists('config_path')) ? config_path('bearychat.php') : base_path('config/bearychat.php');
+            $this->publishes([$from => $to], 'bearychat');
         }
 
         $this->app->singleton('bearychat', function ($app) {
             return (new ClientManager)
-                ->setDefaultName($this->getConfig('default'))
-                ->setClientsDefaults($this->getConfig('clients_defaults'))
-                ->setClientsConfig($this->getConfig('clients'));
+                ->setDefaultName($app['config']->get('bearychat.default'))
+                ->setClientsDefaults($app['config']->get('bearychat.clients_defaults'))
+                ->setClientsConfig($app['config']->get('bearychat.clients'));
         });
 
         $this->app->alias('bearychat', ClientManager::class);
-
-        $this->aliasFacades();
     }
 
     /**
-     * Get the source config path.
+     * Detemines if the application is Lumen.
      *
-     * @return string
+     * @return bool
      */
-    protected function getConfigFromPath()
+    protected function isLumen()
     {
-        return __DIR__.'/config/config.php';
-    }
-
-    /**
-     * Get the config destination path.
-     *
-     * @return string
-     */
-    protected function getConfigToPath()
-    {
-        return $this->isLumen ? base_path('config/bearychat.php') : config_path('bearychat.php');
-    }
-
-    /**
-     * Register facade alias.
-     *
-     * @return void
-     */
-    protected function aliasFacades()
-    {
-        if (class_exists('Illuminate\Foundation\AliasLoader')) {
-            \Illuminate\Foundation\AliasLoader::getInstance()->alias('BearyChat', Facade::class);
-        } else {
-            class_alias(Facade::class, 'BearyChat');
-        }
-    }
-
-    /**
-     * Get the bearychat configuration.
-     *
-     * @param  string  $key
-     * @param  mixed  $default
-     * @return mixed
-     */
-    protected function getConfig($key, $default = null)
-    {
-        $prefix = 'bearychat'.($this->isLaravel4 ? '::' : '.');
-
-        return $this->app['config']->get($prefix.$key, $default);
+        return strpos($this->app->version(), 'Lumen') !== false;
     }
 
     /**
@@ -145,6 +55,6 @@ class ServiceProvider extends LaravelServiceProvider
      */
     public function provides()
     {
-        return ['bearychat'];
+        return ['bearychat', ClientManager::class];
     }
 }